Add travis setup for auto deployment

+++ build_runtime.sh
@@ -1,6 +1,16 @@
 #!/bin/bash
-VERSION=${VERSION:=3.5.1}
-rm -r R/
+
+set -euo pipefail
+
+VERSION=$1
+
+if [ -z "$VERSION" ];
+then
+    echo 'version number required'
+    exit 1
+fi
+
+rm -rf R/
 unzip R-$VERSION.zip -d R/
 rm -r R/doc/manual/
 #remove some libraries to save space
@@ -10,6 +20,5 @@ do
    rm -r R/library/$package/
 done
 chmod -R 755 bootstrap runtime.r R/
-rm runtime.zip
+rm -f runtime.zip
 zip -r runtime.zip runtime.r bootstrap R/
-aws lambda publish-layer-version --layer-name r-runtime --zip-file fileb://runtime.zip
